feat: make DAO routes search-ready
