+++ apps/web/src/app/[locale]/ai-analysis/layout.tsx
@@ -1,1 +1,1 @@
-export { default } from "../../ai-analysis/layout";
+export { default, metadata } from "../../ai-analysis/layout";
