+++ apps/web/src/app/[locale]/profile/layout.tsx
@@ -1,0 +1,1 @@
+export { default, metadata } from "../../profile/layout";
