+++ apps/web/src/app/[locale]/proposals/new/layout.tsx
@@ -1,0 +1,1 @@
+export { default, metadata } from "../../../proposals/new/layout";
