+++ apps/web/src/app/[locale]/proposals/page.tsx
@@ -1,1 +1,1 @@
-export { default } from "../../proposals/page";
+export { default, generateMetadata } from "../../proposals/page";
