+++ apps/web/src/app/ai-analysis/layout.tsx
@@ -1,5 +1,14 @@
 import "../markdown-body.css";
 
+import type { Metadata } from "next";
+
+export const metadata: Metadata = {
+  robots: {
+    index: false,
+    follow: false,
+  },
+};
+
 export default function AiAnalysisLayout({
   children,
 }: {
