+++ apps/web/src/app/page.tsx
@@ -1,15 +1,14 @@
-"use client";
+import { HomeClient } from "./_components/home-client";
+import { DaoPublicSummary } from "./_components/public-route-summary";
+import { getActiveDaoConfig } from "./_server/public-seo";
 
-import { DaoHeader } from "./_components/dao-header";
-import { Overview } from "./_components/overview";
-import { Proposals } from "./_components/proposals";
+export default async function HomePage() {
+  const config = await getActiveDaoConfig();
 
-export default function HomePage() {
   return (
     <div className="flex flex-col gap-[20px] lg:gap-[30px]">
-      <DaoHeader />
-      <Overview />
-      <Proposals />
+      <DaoPublicSummary config={config} />
+      <HomeClient />
     </div>
   );
 }
