+++ apps/web/src/app/profile/layout.tsx
@@ -1,0 +1,12 @@
+import type { Metadata } from "next";
+
+export const metadata: Metadata = {
+  robots: {
+    index: false,
+    follow: false,
+  },
+};
+
+export default function ProfileLayout({ children }: { children: React.ReactNode }) {
+  return children;
+}
