+++ apps/web/src/app/proposals/new/layout.tsx
@@ -1,0 +1,12 @@
+import type { Metadata } from "next";
+
+export const metadata: Metadata = {
+  robots: {
+    index: false,
+    follow: false,
+  },
+};
+
+export default function NewProposalLayout({ children }: { children: React.ReactNode }) {
+  return children;
+}
